+++ config/dev.exs
@@ -1,6 +1,7 @@
 import Config
 
 config :urepo,
+  token: "secret",
   public_key: "public.pem",
   private_key: "private.pem",
   store: {Urepo.Store.Local, bucket: "repo", path: "repo"}