+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Add `CHANGELOG`
 - Add stores documentation
 - Add Distillery releases
+- Add authorization to API requests